Add nonce check for delete txn action

+++ simple-membership/views/payments/admin_all_payment_transactions.php
@@ -27,6 +27,7 @@ if (isset($_REQUEST['action'])) { //Do list table form row action tasks
     if (isset($_REQUEST['action']) && $_REQUEST['action'] == 'delete_txn') { //Delete link was clicked for a row in list table
         $record_id = sanitize_text_field($_REQUEST['id']);
         $record_id = absint($record_id);
+        check_admin_referer('swpm_delete_txn_'.$record_id);
         $payments_list_table->delete_record($record_id);
         $success_msg = '<div id="message" class="updated"><p><strong>';
         $success_msg .= SwpmUtils::_('The selected entry was deleted!');
